+++ collector/src/main/java/com/bitmovin/analytics/data/DeviceInformationProvider.kt
@@ -1,6 +1,7 @@
 package com.bitmovin.analytics.data
 
 import android.content.Context
+import android.content.res.Configuration
 import android.os.Build
 import com.bitmovin.analytics.utils.Util
 
@@ -13,6 +14,11 @@ open class DeviceInformationProvider(val context: Context, val userAgent: String
                 packageName = context.packageName,
                 screenWidth = context?.getResources()?.getDisplayMetrics()?.widthPixels ?: 0,
                 screenHeight = context?.getResources()?.getDisplayMetrics()?.heightPixels ?: 0,
+                screenOrientation = when (context.resources?.configuration?.orientation) {
+                    Configuration.ORIENTATION_PORTRAIT -> DeviceInformation.ScreenOrientation.Portrait
+                    Configuration.ORIENTATION_LANDSCAPE -> DeviceInformation.ScreenOrientation.Landscape
+                    else -> DeviceInformation.ScreenOrientation.Undefined
+                },
                 userAgent = userAgent
         )
     }